simplify normalize

+++ test/src/LibDecimalFloat.decimal.t.sol
@@ -3,12 +3,12 @@ pragma solidity =0.8.25;
 
 import {LibDecimalFloat, ExponentOverflow, NORMALIZED_MAX} from "src/lib/LibDecimalFloat.sol";
 
-import {Test, console} from "forge-std/Test.sol";
+import {Test} from "forge-std/Test.sol";
 
 contract LibDecimalFloatDecimalTest is Test {
     /// Round trip from/to decimal values without precision loss
     function testFixedDecimalRoundTripLossless(uint256 value, uint8 decimals) external pure {
-        value = bound(value, 0, NORMALIZED_MAX);
+        value = bound(value, 0, uint256(NORMALIZED_MAX));
 
         (int256 signedCoefficient, int256 exponent, bool lossless0) =
             LibDecimalFloat.fromFixedDecimalLossy(value, decimals);
